backend/requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.1.1
django-environ==0.9.0
gunicorn==20.1.0
psycopg2==2.9.3
sqlparse==0.4.2
